use pdfrw 0.4 which handles PdfStrings differently
PdfString.decode() now converts the raw tokens to Unicode, rather than to bytes. This is exactly right for metadata and let's us delete some code.
But for text rendering, the Unicode conversion doesn't apply, and we now have to use to_bytes() instead, which does what the old decode() did.

When writing, pdfrw's broken Unicode-string handling in Python 3 is fixed in 0.4 so we can remove the monkeypatching and the conversion in/out
of Latin-1.

See #11.